pyproject.toml
[project]
name = "tracktruck"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
